+++ src/Kerbalism/Modules/Reliability.cs
@@ -20,24 +20,15 @@ namespace KERBALISM
 		[KSPField] public double rated_radiation = 0;                       // rad/h this part can sustain without taking any damage. Only effective with MTBF failures.
 		[KSPField] public double radiation_decay_rate = 1;                  // time to next failure is reduced by (rad/h - rated_radiation) * radiation_decay_rate per second
 
-		// engine only features
-		[KSPField] public double turnon_failure_probability = -1;           // probability of a failure when turned on or staged
-		[KSPField] public double rated_operation_duration = -1;             // failure rate increases dramatically if this is exceeded
-		[KSPField] public int rated_ignitions = -1;                         // failure rate increases dramatically if this is exceeded
-
 		// persistence
 		[KSPField(isPersistant = true)] public bool broken;                 // true if broken
-		[KSPField(isPersistant = true)] public bool critical;               // true if failure can't be repaired
+		[KSPField(isPersistant = true)] public bool critical;               // true if failure requires a more qualified engineer and two repair kits
 		[KSPField(isPersistant = true)] public bool quality;                // true if the component is high-quality
 		[KSPField(isPersistant = true)] public double last = 0.0;           // time of last failure
 		[KSPField(isPersistant = true)] public double next = 0.0;           // time of next failure
 		[KSPField(isPersistant = true)] public double last_inspection = 0.0;   // time of last service
 		[KSPField(isPersistant = true)] public bool needMaintenance = false;// true when component is inspected and about to fail
 		[KSPField(isPersistant = true)] public bool enforce_breakdown = false; // true when the next failure is enforced
-		[KSPField(isPersistant = true)] public bool running = false;        // true when the next failure is enforced
-		[KSPField(isPersistant = true)] public double operation_duration = 0.0; // failure rate increases dramatically if this is exceeded
-		[KSPField(isPersistant = true)] public double fail_duration = 0.0;  // fail when operation_duration exceeds this
-		[KSPField(isPersistant = true)] public int ignitions = 0;           // accumulated ignitions
 		[KSPField(isPersistant = true)] public bool radiator_state_stored;  // true after caching a switch radiator's pre-failure state
 		[KSPField(isPersistant = true)] public bool radiator_was_cooling;   // switch radiator state restored after repair
 
@@ -128,89 +119,6 @@ namespace KERBALISM
 			}
 		}
 
-		/// <summary> Returns true if a failure should be triggered. </summary>
-		protected bool IgnitionCheck()
-		{
-			if (!PreferencesReliability.Instance.engineFailures)
-				return false;
-
-			// don't check for a couple of seconds after the vessel was loaded.
-			// when loading a quicksave with the engines running, the engine state
-			// is off at first which would cost an ignition and possibly trigger a failure
-			if (Time.time < Kerbalism.gameLoadTime + 3)
-				return false;
-
-			ignitions++;
-			vessel.KerbalismData().ResetReliabilityStatus();
-
-			bool fail = false;
-			bool launchpad = vessel.situation == Vessel.Situations.PRELAUNCH || ignitions <= 1 && vessel.situation == Vessel.Situations.LANDED;
-
-			if (turnon_failure_probability > 0)
-			{
-				// q = quality indicator
-				var q = quality ? Settings.QualityScale : 1.0;
-				if (launchpad) q /= 2.5; // the very first ignition is more likely to fail
-
-				q += Lib.Clamp(ignitions - 1, 0.0, 6.0) / 20.0; // subsequent ignition failures become less and less likely, reducing by up to 30%
-
-				if (Lib.RandomDouble() < (turnon_failure_probability * PreferencesReliability.Instance.ignitionFailureChance) / q)
-				{
-					fail = true;
-#if DEBUG_RELIABILITY
-					Lib.DebugLog("Ignition check: " + part.partInfo.title + " ignitions " + ignitions + " turnon failure");
-#endif
-				}
-			}
-
-			// 
-
-			if (rated_ignitions > 0)
-			{
-				int total_ignitions = EffectiveIgnitions(quality, rated_ignitions);
-				if (ignitions >= total_ignitions * 0.9) needMaintenance = true;
-				if (ignitions > total_ignitions)
-				{
-					var q = (quality ? Settings.QualityScale : 1.0) * Lib.RandomDouble();
-					q /= PreferencesReliability.Instance.ignitionFailureChance;
-					q /= (ignitions - total_ignitions); // progressively increase the odds of a failure with every extra ignition
-
-#if DEBUG_RELIABILITY
-					Lib.Log("Reliability: ignition exceeded q=" + q + " ignitions=" + ignitions + " total_ignitions=" + total_ignitions);
-#endif
-
-					if (q < 0.3)
-					{
-						fail = true;
-					}
-				}
-			}
-
-			if (fail)
-			{
-				enforce_breakdown = true;
-				explode = Lib.RandomDouble() < 0.1;
-
-				next = Planetarium.GetUniversalTime() + Lib.RandomDouble() * 2.0;
-
-				var fuelSystemFailureProbability = 0.1;
-				if (launchpad) fuelSystemFailureProbability = 0.5;
-
-				if(Lib.RandomDouble() < fuelSystemFailureProbability)
-				{
-					// broken fuel line -> delayed kaboom
-					explode = true;
-					next += Lib.RandomDouble() * 10 + 4;
-					FlightLogger.fetch?.LogEvent(Local.FlightLogger_Destruction.Format(part.partInfo.title));// <<1>> fuel system leak caused destruction of the engine"
-				}
-				else
-				{
-					FlightLogger.fetch?.LogEvent(Local.FlightLogger_Ignition.Format(part.partInfo.title));// <<1>> failure on ignition"
-				}
-			}
-			return fail;
-		}
-
 		public void Update()
 		{
 			if (Lib.IsFlight())
@@ -239,22 +147,6 @@ namespace KERBALISM
 					}
 					else
 					{
-						if (PreferencesReliability.Instance.engineFailures && (rated_operation_duration > 0 || rated_ignitions > 0))
-						{
-							if (rated_operation_duration > 0)
-							{
-								double effective_duration = EffectiveDuration(quality, rated_operation_duration);
-								Status = Lib.BuildString(Local.Reliability_burnremaining, " ", Lib.HumanReadableDuration(Math.Max(0, effective_duration - operation_duration)));//"remaining burn:"
-							}
-							if (rated_ignitions > 0)
-							{
-								int effective_ignitions = EffectiveIgnitions(quality, rated_ignitions);
-								Status = Lib.BuildString(Status,
-									(string.IsNullOrEmpty(Status) ? "" : ", "),
-									Local.Reliability_ignitions, " ", Math.Max(0, effective_ignitions - ignitions).ToString());//"ignitions:"
-							}
-						}
-
 						if (rated_radiation > 0)
 						{
 							var rated = quality ? rated_radiation * Settings.QualityScale : rated_radiation;
@@ -269,15 +161,13 @@ namespace KERBALISM
 					if (string.IsNullOrEmpty(Status)) Status = Local.Generic_NOMINAL;//"nominal"
 
 					Events["Inspect"].active = !broken && !needMaintenance;
-					Events["Repair"].active = repair_cs && (broken || needMaintenance) && !critical;
+					Events["Repair"].active = repair_cs && (broken || needMaintenance);
 
 					if (needMaintenance)
 					{
 						Events["Repair"].guiName = Local.Reliability_Service.Format("<b>" + LocalizeTitle(title) + "</b>");//Lib.BuildString("Service <<1>>")
 					}
 				}
-
-				RunningCheck();
 
 				// if it has failed, trigger malfunction
 				var now = Planetarium.GetUniversalTime();
@@ -306,23 +196,6 @@ namespace KERBALISM
 						Status = Lib.BuildString(Status,
 							(string.IsNullOrEmpty(Status) ? "" : ", "),
 							Local.Reliability_MTBF + " ", Lib.HumanReadableDuration(effective_mtbf));//"MTBF:"
-					}
-
-					if (rated_operation_duration > 0 && PreferencesReliability.Instance.engineFailures)
-					{
-						double effective_duration = EffectiveDuration(quality, rated_operation_duration);
-						Status = Lib.BuildString(Status,
-							(string.IsNullOrEmpty(Status) ? "" : ", "),
-							Local.Reliability_Burntime + " ",//"Burn time:
-							Lib.HumanReadableDuration(effective_duration));
-					}
-
-					if (rated_ignitions > 0 && PreferencesReliability.Instance.engineFailures)
-					{
-						int effective_ignitions = EffectiveIgnitions(quality, rated_ignitions);
-						Status = Lib.BuildString(Status,
-							(string.IsNullOrEmpty(Status) ? "" : ", "),
-							Local.Reliability_ignitions + " ", effective_ignitions.ToString());//"ignitions:
 					}
 
 					if (rated_radiation > 0 && PreferencesReliability.Instance.mtbfFailures)
@@ -361,101 +234,6 @@ namespace KERBALISM
 				var decay = RadiationDecay(quality, vessel.KerbalismData().EnvRadiation, Kerbalism.elapsed_s, rated_radiation, radiation_decay_rate);
 				next -= decay;
 			}
-		}
-
-		protected double nextRunningCheck = 0.0;
-		protected double lastRunningCheck = 0.0;
-
-		/// <summary>
-		/// This checks burn time and ignition failures, which is intended for engines only.
-		/// Since engines don't run on on unloaded vessels in KSP, this is only implemented
-		/// for loaded vessels.
-		/// </summary>
-		protected void RunningCheck()
-		{
-			if (!PreferencesReliability.Instance.engineFailures) return;
-
-			// don't count engine running time during time warp (see https://github.com/Kerbalism/Kerbalism/pull/646)
-			if (TimeWarp.WarpMode == TimeWarp.Modes.HIGH && TimeWarp.CurrentRate > 1)
-			{
-				// forget that the engine was running when we're warping. otherwise we might consume engine time
-				// during warp when it actually didn't run because of KSP physics
-				lastRunningCheck = 0;
-				return;
-			}
-
-			if (broken || enforce_breakdown || turnon_failure_probability <= 0 && rated_operation_duration <= 0) return;
-			double now = Planetarium.GetUniversalTime();
-			if (now < nextRunningCheck) return;
-			nextRunningCheck = now + 0.5; // twice a second is fast enough for a smooth countdown in the PAW
-
-			if (!running)
-			{
-				if (IsRunning())
-				{
-					running = true;
-					if (IgnitionCheck())
-						Break();
-				}
-			}
-			else
-			{
-				running = IsRunning();
-			}
-
-			if (running && rated_operation_duration > 1 && lastRunningCheck > 0)
-			{
-				var duration = now - lastRunningCheck;
-				operation_duration += duration;
-				vessel.KerbalismData().ResetReliabilityStatus();
-
-				if (fail_duration <= 0)
-				{
-					// see https://www.desmos.com/calculator/l2sdoauiyw
-					// these values will give the following result:
-					// 0% failure at 40% of the operation duration
-					// 1.68% failure at 100% of the operation duration
-					// 100% failure at 140% of the operation duration
-
-					int r = 8; // the random number exponent
-					var g = 0.4; // guarantee factor
-
-					// calculate a random point on which the engine will fail
-
-					var f = rated_operation_duration;
-					if (quality) f *= Settings.QualityScale;
-
-					// extend engine burn duration by preference value 
-					f /= PreferencesReliability.Instance.engineOperationFailureChance;
-
-					// random^r so we get an exponentially increasing failure probability
-					var p = Math.Pow(Lib.RandomDouble(), r);
-
-					// 1-p turns the probability of failure into one of non-failure
-					p = 1 - p;
-
-					// 35% guaranteed burn duration
-					var guaranteed_operation = f * g;
-
-					fail_duration = guaranteed_operation + f * p;
-#if DEBUG_RELIABILITY
-					Lib.Log(part.partInfo.title + " will fail after " + Lib.HumanReadableDuration(fail_duration) + " burn time");
-#endif
-				}
-
-				if (fail_duration < operation_duration)
-				{
-					next = now;
-					enforce_breakdown = true;
-					explode = Lib.RandomDouble() < 0.2;
-#if DEBUG_RELIABILITY
-					Lib.Log("Reliability: " + part.partInfo.title + " fails because of material fatigue");
-#endif
-					FlightLogger.fetch?.LogEvent(Local.FlightLogger_MaterialFatigue.Format(part.partInfo.title));// <<1>> failed because of material fatigue"
-				}
-			}
-
-			lastRunningCheck = now;
 		}
 
 		public static void BackgroundUpdate(Vessel v, ProtoPartSnapshot p, ProtoPartModuleSnapshot m, Reliability reliability, double elapsed_s)
@@ -533,9 +311,6 @@ namespace KERBALISM
 			double time_k = (Planetarium.GetUniversalTime() - last) / (next - last);
 			needMaintenance = mtbf > 0 && time_k > 0.35;
 
-			if (rated_ignitions > 0 && ignitions >= Math.Ceiling(EffectiveIgnitions(quality, rated_ignitions) * 0.4)) needMaintenance = true;
-			if (rated_operation_duration > 0 && operation_duration >= EffectiveDuration(quality, rated_operation_duration) * 0.4) needMaintenance = true;
-
 			v.KerbalismData().ResetReliabilityStatus();
 
 			// notify user
@@ -570,7 +345,8 @@ namespace KERBALISM
 			if (v == null) return;
 
 			// check trait
-			if (!repair_cs.Check(v))
+			CrewSpecs requiredRepairCrew = critical ? repair_cs.ElevatedForCritical() : repair_cs;
+			if (!requiredRepairCrew.Check(v))
 			{
 				Message.Post
 				(
@@ -580,7 +356,7 @@ namespace KERBALISM
 					Local.Reliability_MessagePost12,//"I will not even know where to start"
 					Local.Reliability_MessagePost13//"I'm afraid I can't do that"
 				  ),
-				  repair_cs.Warning()
+				  requiredRepairCrew.Warning()
 				);
 				return;
 			}
@@ -591,52 +367,18 @@ namespace KERBALISM
 			// reset times
 			last = 0.0;
 			next = 0.0;
-			lastRunningCheck = 0;
 			last_inspection = Planetarium.GetUniversalTime();
 
-			operation_duration = 0;
-			ignitions = 0;
-			fail_duration = 0;
 			vessel.KerbalismData().ResetReliabilityStatus();
 
 			if (broken)
 			{
-				if (PreferencesReliability.Instance.requireRepairKits) // REQUIRE REPAIR KIT - Settings Option
-				{
-					int repairKits = 0;
-					KerbalEVA kerbalEVA = v.evaController;
-					if (kerbalEVA.ModuleInventoryPartReference != null && v.isEVA)
-					{
-						foreach (StoredPart storedPart in kerbalEVA.ModuleInventoryPartReference.storedParts.Values)
-						{
-							// Note : the "evaRepairKit" string is hardcoded in the KSP source
-							if (storedPart.partName == "evaRepairKit")
-							{
-								repairKits++;
-							}
-						}
-					}
-					if (repairKits <= 0)
-					{
-						Message.Post
-						(
-						  Local.Reliability_MessagePost30.Format("<b>" + LocalizeTitle(title) + "</b>"),//Lib.BuildString("<<1>> needs a repair kit")
-						  Lib.TextVariant
-						  (
-							Local.Reliability_MessagePost31,//"Did I forget something."
-							Local.Reliability_MessagePost32//"Oh crap..."
-						  )
-						);
-						return;
-					}
-					else
-					{
-						kerbalEVA.ModuleInventoryPartReference.RemoveNPartsFromInventory("evaRepairKit", 1, true);
-					}
-				}
+				int repairKitCost = critical ? 2 : 1;
+				if (!ConsumeRepairKits(v, LocalizeTitle(title), repairKitCost)) return;
 
 				// flag as not broken
 				broken = false;
+				critical = false;
 
 				// re-enable module
 				foreach (PartModule m in modules)
@@ -679,6 +421,39 @@ namespace KERBALISM
 				  )
 				);
 			}
+		}
+
+		public static bool ConsumeRepairKits(Vessel v, string localizedTitle, int amount)
+		{
+			if (!PreferencesReliability.Instance.requireRepairKits) return true;
+
+			int repairKits = 0;
+			KerbalEVA kerbalEVA = v.evaController;
+			if (v.isEVA && kerbalEVA != null && kerbalEVA.ModuleInventoryPartReference != null)
+			{
+				foreach (StoredPart storedPart in kerbalEVA.ModuleInventoryPartReference.storedParts.Values)
+				{
+					// Note: the "evaRepairKit" string is hardcoded in the KSP source.
+					if (storedPart.partName == "evaRepairKit") repairKits += storedPart.quantity;
+				}
+			}
+
+			if (repairKits < amount)
+			{
+				Message.Post
+				(
+				  Local.Reliability_MessagePost30.Format("<b>" + localizedTitle + "</b>"),//Lib.BuildString("<<1>> needs a repair kit")
+				  Lib.TextVariant
+				  (
+					Local.Reliability_MessagePost31,//"Did I forget something."
+					Local.Reliability_MessagePost32//"Oh crap..."
+				  )
+				);
+				return false;
+			}
+
+			kerbalEVA.ModuleInventoryPartReference.RemoveNPartsFromInventory("evaRepairKit", amount, true);
+			return true;
 		}
 
 #if DEBUG_RELIABILITY
@@ -915,9 +690,6 @@ namespace KERBALISM
 			specs.Add(string.Empty);
 			specs.Add("<color=#00ffff>"+Local.Reliability_info3 +"</color>");//Standard quality
 			if(mtbf > 0) specs.Add(Local.Reliability_info4, Lib.HumanReadableDuration(EffectiveMTBF(false, mtbf)));//"MTBF"
-			if (turnon_failure_probability > 0) specs.Add(Local.Reliability_info5, Lib.HumanReadablePerc(turnon_failure_probability, "F1"));//"Ignition failures"
-			if (rated_operation_duration > 0) specs.Add(Local.Reliability_info6, Lib.HumanReadableDuration(EffectiveDuration(false, rated_operation_duration)));//"Rated burn duration"
-			if (rated_ignitions > 0) specs.Add(Local.Reliability_info7, EffectiveIgnitions(false, rated_ignitions).ToString());//"Rated ignitions"
 			if (mtbf > 0 && rated_radiation > 0) specs.Add(Local.Reliability_info8, Lib.HumanReadableRadiation(rated_radiation / 3600.0));//"Radiation rating"
 
 			specs.Add(string.Empty);
@@ -925,9 +697,6 @@ namespace KERBALISM
 			if (extra_cost > double.Epsilon) specs.Add(Local.Reliability_info10, Lib.HumanReadableCost(extra_cost * part.partInfo.cost));//"Extra cost"
 			if (extra_mass > double.Epsilon) specs.Add(Local.Reliability_info11, Lib.HumanReadableMass(extra_mass * part.partInfo.partPrefab.mass));//"Extra mass"
 			if (mtbf > 0) specs.Add(Local.Reliability_info4, Lib.HumanReadableDuration(EffectiveMTBF(true, mtbf)));//"MTBF"
-			if (turnon_failure_probability > 0) specs.Add(Local.Reliability_info5, Lib.HumanReadablePerc(turnon_failure_probability / Settings.QualityScale, "F1"));//"Ignition failures"
-			if (rated_operation_duration > 0) specs.Add(Local.Reliability_info6, Lib.HumanReadableDuration(EffectiveDuration(true, rated_operation_duration)));//"Rated burn duration"
-			if (rated_ignitions > 0) specs.Add(Local.Reliability_info7, EffectiveIgnitions(true, rated_ignitions).ToString());//"Rated ignitions"
 			if (mtbf > 0 && rated_radiation > 0) specs.Add(Local.Reliability_info8, Lib.HumanReadableRadiation(Settings.QualityScale * rated_radiation / 3600.0));//"Radiation rating"
 
 			return specs;
@@ -948,34 +717,6 @@ namespace KERBALISM
 		public float GetModuleMass(float defaultMass, ModifierStagingSituation sit) { return quality ? (float)extra_mass * part.partInfo.partPrefab.mass : 0.0f; }
 		public ModifierChangeWhen GetModuleCostChangeWhen() { return ModifierChangeWhen.CONSTANTLY; }
 		public ModifierChangeWhen GetModuleMassChangeWhen() { return ModifierChangeWhen.CONSTANTLY; }
-
-		protected bool IsRunning()
-		{
-			if (type.StartsWith("ModuleEngines", StringComparison.Ordinal))
-			{
-				foreach (PartModule m in modules)
-				{
-					var e = m as ModuleEngines;
-					return e.currentThrottle > 0 && e.EngineIgnited && e.resultingThrust > 0;
-				}
-				return false;
-			}
-
-			switch (type)
-			{
-				case "ProcessController":
-					foreach (PartModule m in modules)
-						return (m as ProcessController).running;
-					return false;
-
-				case "ModuleLight":
-					foreach (PartModule m in modules)
-						return (m as ModuleLight).isOn;
-					return false;
-			}
-
-			return false;
-		}
 
 		void SetAlternatorsEnabled(bool enabled)
 		{
@@ -1343,7 +1084,7 @@ namespace KERBALISM
 					(
 					  Severity.danger,
 					  Local.Reliability_MessagePost26.Format("<b>" + title + "</b>", "<b>" + v.vesselName + "</b>"),//Lib.BuildString(<<1>> failed on <<2>>)
-					  Local.Reliability_MessagePost27//"It is gone for good"
+					  Local.Reliability_MessagePost27//"It can still be repaired, but requires greater expertise"
 					);
 				}
 			}